Sync darwinia-common service part
