+++ cli/src/command.rs
@@ -166,16 +166,18 @@ pub fn run() -> CliResult<()> {
 			if chain_spec.is_crab() {
 				runner.async_run(|mut config| {
 					let (client, _, import_queue, task_manager) =
-						crab_service::new_chain_ops::<CrabRuntimeApi, CrabExecutor>(&mut config)?;
+						darwinia_node_service::new_chain_ops::<CrabRuntimeApi, CrabExecutor>(
+							&mut config,
+						)?;
 
 					Ok((cmd.run(client, import_queue), task_manager))
 				})
 			} else {
 				runner.async_run(|mut config| {
-					let (client, _, import_queue, task_manager) = darwinia_service::new_chain_ops::<
-						DarwiniaRuntimeApi,
-						DarwiniaExecutor,
-					>(&mut config)?;
+					let (client, _, import_queue, task_manager) =
+						darwinia_node_service::new_chain_ops::<DarwiniaRuntimeApi, DarwiniaExecutor>(
+							&mut config,
+						)?;
 
 					Ok((cmd.run(client, import_queue), task_manager))
 				})
@@ -189,14 +191,16 @@ pub fn run() -> CliResult<()> {
 
 			if chain_spec.is_crab() {
 				runner.async_run(|mut config| {
-					let (client, _, _, task_manager) =
-						crab_service::new_chain_ops::<CrabRuntimeApi, CrabExecutor>(&mut config)?;
+					let (client, _, _, task_manager) = darwinia_node_service::new_chain_ops::<
+						CrabRuntimeApi,
+						CrabExecutor,
+					>(&mut config)?;
 
 					Ok((cmd.run(client, config.database), task_manager))
 				})
 			} else {
 				runner.async_run(|mut config| {
-					let (client, _, _, task_manager) = darwinia_service::new_chain_ops::<
+					let (client, _, _, task_manager) = darwinia_node_service::new_chain_ops::<
 						DarwiniaRuntimeApi,
 						DarwiniaExecutor,
 					>(&mut config)?;
@@ -213,14 +217,16 @@ pub fn run() -> CliResult<()> {
 
 			if chain_spec.is_crab() {
 				runner.async_run(|mut config| {
-					let (client, _, _, task_manager) =
-						crab_service::new_chain_ops::<CrabRuntimeApi, CrabExecutor>(&mut config)?;
+					let (client, _, _, task_manager) = darwinia_node_service::new_chain_ops::<
+						CrabRuntimeApi,
+						CrabExecutor,
+					>(&mut config)?;
 
 					Ok((cmd.run(client, config.chain_spec), task_manager))
 				})
 			} else {
 				runner.async_run(|mut config| {
-					let (client, _, _, task_manager) = darwinia_service::new_chain_ops::<
+					let (client, _, _, task_manager) = darwinia_node_service::new_chain_ops::<
 						DarwiniaRuntimeApi,
 						DarwiniaExecutor,
 					>(&mut config)?;
@@ -238,16 +244,18 @@ pub fn run() -> CliResult<()> {
 			if chain_spec.is_crab() {
 				runner.async_run(|mut config| {
 					let (client, _, import_queue, task_manager) =
-						crab_service::new_chain_ops::<CrabRuntimeApi, CrabExecutor>(&mut config)?;
+						darwinia_node_service::new_chain_ops::<CrabRuntimeApi, CrabExecutor>(
+							&mut config,
+						)?;
 
 					Ok((cmd.run(client, import_queue), task_manager))
 				})
 			} else {
 				runner.async_run(|mut config| {
-					let (client, _, import_queue, task_manager) = darwinia_service::new_chain_ops::<
-						DarwiniaRuntimeApi,
-						DarwiniaExecutor,
-					>(&mut config)?;
+					let (client, _, import_queue, task_manager) =
+						darwinia_node_service::new_chain_ops::<DarwiniaRuntimeApi, DarwiniaExecutor>(
+							&mut config,
+						)?;
 
 					Ok((cmd.run(client, import_queue), task_manager))
 				})
@@ -282,14 +290,16 @@ pub fn run() -> CliResult<()> {
 
 			if chain_spec.is_crab() {
 				runner.async_run(|mut config| {
-					let (client, backend, _, task_manager) =
-						crab_service::new_chain_ops::<CrabRuntimeApi, CrabExecutor>(&mut config)?;
+					let (client, backend, _, task_manager) = darwinia_node_service::new_chain_ops::<
+						CrabRuntimeApi,
+						CrabExecutor,
+					>(&mut config)?;
 
 					Ok((cmd.run(client, backend), task_manager))
 				})
 			} else {
 				runner.async_run(|mut config| {
-					let (client, backend, _, task_manager) = darwinia_service::new_chain_ops::<
+					let (client, backend, _, task_manager) = darwinia_node_service::new_chain_ops::<
 						DarwiniaRuntimeApi,
 						DarwiniaExecutor,
 					>(&mut config)?;
